move all types to types.ts
