+++ packages/vm/src/index.ts
@@ -1,2 +1,4 @@
-export { VM as default } from './vm'
+export * as default from './vm'
 export { VMOpts } from './types'
+export * as EEI from './eei/eei'
+export * as Bloom from './bloom/index'
