+++ packages/vm/src/runTx.ts
@@ -3,7 +3,6 @@ import { Address, KECCAK256_NULL, toBuffer, short } from '@ethereumjs/util'
 import { Block } from '@ethereumjs/block'
 import { ConsensusType, Hardfork } from '@ethereumjs/common'
 import {
-  AccessList,
   AccessListItem,
   AccessListEIP2930Transaction,
   FeeMarketEIP1559Transaction,
@@ -11,107 +10,20 @@ import {
   TypedTransaction,
   Capability,
 } from '@ethereumjs/tx'
-import VM from './index'
+import VM from './vm'
 import Bloom from './bloom'
-import { EVMResult } from '@ethereumjs/evm'
 import type {
   TxReceipt,
   BaseTxReceipt,
   PreByzantiumTxReceipt,
   PostByzantiumTxReceipt,
+  RunTxOpts,
+  RunTxResult,
+  AfterTxEvent,
 } from './types'
 
 const debug = createDebugLogger('vm:tx')
 const debugGas = createDebugLogger('vm:tx:gas')
-
-/**
- * Options for the `runTx` method.
- */
-export interface RunTxOpts {
-  /**
-   * The `@ethereumjs/block` the `tx` belongs to.
-   * If omitted, a default blank block will be used.
-   */
-  block?: Block
-  /**
-   * An `@ethereumjs/tx` to run
-   */
-  tx: TypedTransaction
-  /**
-   * If true, skips the nonce check
-   */
-  skipNonce?: boolean
-  /**
-   * Skip balance checks if true. Adds transaction cost to balance to ensure execution doesn't fail.
-   */
-  skipBalance?: boolean
-
-  /**
-   * If true, skips the validation of the tx's gas limit
-   * against the block's gas limit.
-   */
-  skipBlockGasLimitValidation?: boolean
-
-  /**
-   * If true, adds a generated EIP-2930 access list
-   * to the `RunTxResult` returned.
-   *
-   * Option works with all tx types. EIP-2929 needs to
-   * be activated (included in `berlin` HF).
-   *
-   * Note: if this option is used with a custom {@link StateManager} implementation
-   * {@link StateManager.generateAccessList} must be implemented.
-   */
-  reportAccessList?: boolean
-
-  /**
-   * To obtain an accurate tx receipt input the block gas used up until this tx.
-   */
-  blockGasUsed?: bigint
-}
-
-/**
- * Execution result of a transaction
- */
-export interface RunTxResult extends EVMResult {
-  /**
-   * Bloom filter resulted from transaction
-   */
-  bloom: Bloom
-
-  /**
-   * The amount of ether used by this transaction
-   */
-  amountSpent: bigint
-
-  /**
-   * The tx receipt
-   */
-  receipt: TxReceipt
-
-  /**
-   * The amount of gas used in this transaction, which is paid for
-   * This contains the gas units t
-   */
-  totalGasSpent: bigint
-
-  /**
-   * The amount of gas as that was refunded during the transaction (i.e. `gasUsed = totalGasConsumed - gasRefund`)
-   */
-  gasRefund: bigint
-
-  /**
-   * EIP-2930 access list generated for the tx (see `reportAccessList` option)
-   */
-  accessList?: AccessList
-}
-
-export interface AfterTxEvent extends RunTxResult {
-  /**
-   * The transaction which just got finished
-   */
-  transaction: TypedTransaction
-}
 
 /**
  * @ignore
